+++ test/unit-tests/plugin/impl/  cdk-build-task-plugin.test.ts
@@ -1,4 +1,4 @@
-import { CdkBuildTaskPlugin, ICdkTask } from "~plugin/impl/cdk-build-task-plugin";
+import { CdkBuildTaskPlugin, ICdkTask, ICdkCommandArgs } from "~plugin/impl/cdk-build-task-plugin";
 import { ChildProcessUtility } from "~util/child-process-util";
 import { IPluginBinding, PluginBinder } from "~plugin/plugin-binder";
 import { ICfnSubExpression, ICfnGetAttExpression, ICfnRefExpression, ICfnCopyValue } from "~core/cfn-expression";
@@ -6,6 +6,7 @@ import { TemplateRoot } from "~parser/parser";
 import { PersistedState } from "~state/persisted-state";
 import { TestTemplates } from "../../../../test/unit-tests/test-templates";
 import { AwsUtil } from "~util/aws-util";
+import { IPerformTasksCommandArgs } from "~commands/index";
 
 describe('when creating cdk plugin', () => {
     let plugin: CdkBuildTaskPlugin;
@@ -46,6 +47,37 @@ describe('when creating cdk plugin', () => {
         expect(commandArgs.customDeployCommand).toBeUndefined();
     });
 
+});
+
+describe('when validating task', () => {
+    let plugin: CdkBuildTaskPlugin;
+    let commandArgs: ICdkCommandArgs;
+
+    beforeEach(() => {
+        plugin = new CdkBuildTaskPlugin();
+        commandArgs = plugin.convertToCommandArgs( {
+            FilePath: './tasks.yaml',
+            Type: 'cdk',
+            MaxConcurrentTasks: 1,
+            FailedTaskTolerance: 4,
+            LogicalName: 'test-task',
+            Path: './',
+            TaskRoleName: 'TaskRole',
+            OrganizationBinding: { IncludeMasterAccount: true}},
+            { organizationFile: './organization.yml'} as any);
+    });
+
+    test('CustomDeployCommand with Sub Expression throws', () => {
+        (commandArgs as any).customDeployCommand = { 'Fn::Sub': 'expression xyz' } as ICfnSubExpression;
+        expect( ()=> { plugin.validateCommandArgs(commandArgs) }).toThrowError(/xyz/);
+        expect( ()=> { plugin.validateCommandArgs(commandArgs) }).toThrowError(/CustomDeployCommand/);
+    });
+
+    test('CustomRemoveCommand with Sub Expression throws', () => {
+        (commandArgs as any).customRemoveCommand = { 'Fn::Sub': 'expression xyz' } as ICfnSubExpression;
+        expect( ()=> { plugin.validateCommandArgs(commandArgs) }).toThrowError(/xyz/);
+        expect( ()=> { plugin.validateCommandArgs(commandArgs) }).toThrowError(/CustomRemoveCommand/);
+    });
 });
 
 describe('when resolving attribute expressions on update', () => {