+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to aws organization formation will be documented in this file.
 
+
+**unreleased**
+- Fixed bug with variables in CustomDeployCommand for SLS tasks.
+- Improved validation and error messages for CustomDeployCommand and CustomRemoveCommand.
+
 **version 0.9.7**
 - Increased buffer size for tasks ran on the console (update-sls and update-cdk)
 - Fixed bug with OU structures that go three levels deep and/or multiple times the same ou name